requirements.txt
# Core LangChain and LangGraph
langchain==0.2.11
langgraph==0.2.3
langchain-community==0.2.10
langchain-openai==0.1.19

# Document processing
pypdf2==3.0.1
pypdf
pdfplumber==0.11.0
PyMuPDF>=1.23.0
python-docx==1.1.0

# Vector store and embeddings
chromadb==0.5.3
sentence-transformers==2.7.0

# Web interface
streamlit==1.28.0
gradio==4.29.0

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# HTTP and API
httpx==0.27.0
requests==2.32.3

# Utilities
numpy==1.24.4
pandas==2.0.3
tqdm==4.66.4
rich==13.7.1

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==8.2.2
black==24.4.2
isort==5.13.2
